+++ grafana_wtf/core.py
@@ -3,11 +3,14 @@
 # License: GNU Affero General Public License, Version 3
 import colored
 import logging
+import asyncio
+import requests
 import requests_cache
 from tqdm import tqdm
 from munch import Munch, munchify
-from urllib.parse import urlparse, urljoin
 from collections import OrderedDict
+from urllib.parse import urlparse, urljoin
+from concurrent.futures.thread import ThreadPoolExecutor
 
 from grafana_api.grafana_api import GrafanaClientError, GrafanaUnauthorizedError
 from grafana_api.grafana_face import GrafanaFace
@@ -27,6 +30,13 @@ class GrafanaSearch:
 
         self.finder = JsonPathFinder()
 
+        self.taqadum = None
+        self.concurrency = 5
+
+        self.debug = log.getEffectiveLevel() == logging.DEBUG
+        if not self.debug:
+            self.taqadum = tqdm()
+
     def enable_cache(self, expire_after=300, drop_cache=False):
         if expire_after is None:
             log.info(f'Setting up response cache to never expire (infinite caching)')
@@ -38,6 +48,9 @@ class GrafanaSearch:
             requests_cache.clear()
 
         return self
+
+    def enable_concurrency(self, concurrency):
+        self.concurrency = concurrency
 
     def setup(self):
         url = urlparse(self.grafana_url)
@@ -55,6 +68,14 @@ class GrafanaSearch:
         self.grafana = GrafanaFace(
             auth, protocol=url.scheme,
             host=url.hostname, port=url.port, url_path_prefix=url.path)
+
+        # Configure a larger HTTP request pool.
+        # Todo: Review the pool settings and eventually adjust according to concurrency level or other parameters.
+        # https://urllib3.readthedocs.io/en/latest/advanced-usage.html#customizing-pool-behavior
+        # https://laike9m.com/blog/requests-secret-pool_connections-and-pool_maxsize,89/
+        adapter = requests.adapters.HTTPAdapter(pool_connections=100, pool_maxsize=100, max_retries=5, pool_block=True)
+        self.grafana.api.s.mount('http://', adapter)
+        self.grafana.api.s.mount('https://', adapter)
 
         return self
 
@@ -147,6 +168,7 @@ class GrafanaSearch:
         try:
             self.data.dashboard_list = self.grafana.search.search_dashboards(limit=5000)
             log.info('Found {} dashboards'.format(len(self.data.dashboard_list)))
+
         except GrafanaClientError as ex:
             message = '{name}: {ex}'.format(name=ex.__class__.__name__, ex=ex)
             message = colored.stylize(message, colored.fg("red") + colored.attr("bold"))
@@ -154,11 +176,58 @@ class GrafanaSearch:
             if isinstance(ex, GrafanaUnauthorizedError):
                 log.error(self.get_red_message('Please use --grafana-token or GRAFANA_TOKEN '
                                                'for authenticating with Grafana'))
+            return
 
-        log.info('Fetching dashboards')
-        for dashboard_info in tqdm(self.data.dashboard_list):
-            dashboard = self.grafana.dashboard.get_dashboard(dashboard_info['uid'])
-            self.data.dashboards.append(dashboard)
+        if self.taqadum is not None:
+            self.taqadum.total = len(self.data.dashboard_list)
+
+        if self.concurrency is None or self.concurrency <= 1:
+            self.fetch_dashboards()
+        else:
+            self.fetch_dashboards_parallel()
+
+    def fetch_dashboard(self, dashboard_info):
+        log.debug(f'Fetching dashboard "{dashboard_info["title"]}" ({dashboard_info["uid"]})')
+        dashboard = self.grafana.dashboard.get_dashboard(dashboard_info['uid'])
+        self.data.dashboards.append(dashboard)
+        if self.taqadum is not None:
+            self.taqadum.update(1)
+
+    def fetch_dashboards(self):
+        log.info('Fetching dashboards one by one')
+        results = self.data.dashboard_list
+        #if not self.debug:
+        #    results = tqdm(results)
+        for dashboard_info in results:
+            self.fetch_dashboard(dashboard_info)
+
+    def fetch_dashboards_parallel(self):
+        # https://hackernoon.com/how-to-run-asynchronous-web-requests-in-parallel-with-python-3-5-without-aiohttp-264dc0f8546
+        loop = asyncio.get_event_loop()
+        future = asyncio.ensure_future(self.execute_parallel())
+        loop.run_until_complete(future)
+
+    async def execute_parallel(self):
+        # https://hackernoon.com/how-to-run-asynchronous-web-requests-in-parallel-with-python-3-5-without-aiohttp-264dc0f8546
+        log.info(f'Fetching dashboards in parallel with {self.concurrency} concurrent requests')
+        with ThreadPoolExecutor(max_workers=self.concurrency) as executor:
+
+            # Set any session parameters here before calling `fetch`
+            loop = asyncio.get_event_loop()
+            # START_TIME = default_timer()
+
+            tasks = []
+            for dashboard_info in self.data.dashboard_list:
+                task = loop.run_in_executor(
+                    executor,
+                    self.fetch_dashboard,
+                    dashboard_info
+                )
+                tasks.append(task)
+
+            # Currently, we are not interested in the responses.
+            #for response in await asyncio.gather(*tasks):
+            #    pass
 
     def get_dashboard_versions(self, dashboard_id):
         """
@@ -170,3 +239,7 @@ class GrafanaSearch:
         get_dashboard_versions_path = '/dashboards/id/%s/versions' % dashboard_id
         r = self.grafana.dashboard.api.GET(get_dashboard_versions_path)
         return r
+
+    def __del__(self):
+        if self.taqadum:
+            self.taqadum.close()